requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-log-request-id==2.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
psycopg2==2.9.9
coverage==7.4.4
setuptools==69.2.0 # Github action errors
Faker==24.4.0
django-filter==24.2
requests==2.31.0
pusher==3.3.2